# backend/openapi.yaml
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    UserPreferencesOut:
      properties:
        created_at:
          format: date-time
          title: Created At
          type: string
        daily_study_time:
          default: 30
          title: Daily Study Time
          type: integer
        id:
          title: Id
          type: integer
        learning_goals:
          anyOf:
          - type: string
          - type: 'null'
          title: Learning Goals
        notifications_enabled:
          default: true
          title: Notifications Enabled
          type: boolean
        preferred_categories:
          anyOf:
          - type: string
          - type: 'null'
          title: Preferred Categories
        skill_level:
          default: beginner
          title: Skill Level
          type: string
        user_id:
          title: User Id
          type: integer
      required:
      - id
      - user_id
      - created_at
      title: UserPreferencesOut
      type: object
    UserPreferencesUpdate:
      properties:
        daily_study_time:
          anyOf:
          - type: integer
          - type: 'null'
          title: Daily Study Time
        learning_goals:
          anyOf:
          - type: string
          - type: 'null'
          title: Learning Goals
        notifications_enabled:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Notifications Enabled
        preferred_categories:
          anyOf:
          - type: string
          - type: 'null'
          title: Preferred Categories
        skill_level:
          anyOf:
          - type: string
          - type: 'null'
          title: Skill Level
      title: UserPreferencesUpdate
      type: object
    UserProgress:
      properties:
        completed:
          default: false
          title: Completed
          type: boolean
        course_id:
          title: Course Id
          type: integer
        created_at:
          format: date-time
          title: Created At
          type: string
        id:
          title: Id
          type: integer
        last_accessed:
          format: date-time
          title: Last Accessed
          type: string
        lesson_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Lesson Id
        progress_percentage:
          default: 0.0
          title: Progress Percentage
          type: number
        time_spent_minutes:
          default: 0
          title: Time Spent Minutes
          type: integer
        user_id:
          title: User Id
          type: integer
      required:
      - id
      - user_id
      - course_id
      - lesson_id
      - last_accessed
      - created_at
      title: UserProgress
      type: object
    UserProgressUpdate:
      properties:
        completed:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Completed
        progress_percentage:
          anyOf:
          - type: number
          - type: 'null'
          title: Progress Percentage
        time_spent_minutes:
          anyOf:
          - type: integer
          - type: 'null'
          title: Time Spent Minutes
      title: UserProgressUpdate
      type: object
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
  securitySchemes:
    OAuth2PasswordBearer:
      flows:
        password:
          scopes: {}
          tokenUrl: /api/auth/login
      type: oauth2
info:
  title: Gyanvruksh API
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /api/progress/courses/{course_id}:
    get:
      description: Get user's progress for a specific course
      operationId: get_course_progress_api_progress_courses__course_id__get
      parameters:
      - in: path
        name: course_id
        required: true
        schema:
          title: Course Id
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProgress'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - OAuth2PasswordBearer: []
      summary: Get Course Progress
      tags:
      - progress
  /api/progress/courses/{course_id}/lessons/{lesson_id}:
    post:
      description: Update or create progress for a specific lesson
      operationId: update_lesson_progress_api_progress_courses__course_id__lessons__lesson_id__post
      parameters:
      - in: path
        name: course_id
        required: true
        schema:
          title: Course Id
          type: integer
      - in: path
        name: lesson_id
        required: true
        schema:
          title: Lesson Id
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserProgressUpdate'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProgress'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - OAuth2PasswordBearer: []
      summary: Update Lesson Progress
      tags:
      - progress
  /api/progress/detailed:
    get:
      description: Get detailed progress information including analytics
      operationId: get_detailed_progress_api_progress_detailed_get
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
      security:
      - OAuth2PasswordBearer: []
      summary: Get Detailed Progress
      tags:
      - progress
  /api/progress/gamification:
    get:
      description: Get gamification data including streaks, points, badges
      operationId: get_gamification_data_api_progress_gamification_get
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
      security:
      - OAuth2PasswordBearer: []
      summary: Get Gamification Data
      tags:
      - progress
  /api/progress/overall:
    get:
      description: Get overall user progress across all courses
      operationId: get_overall_progress_api_progress_overall_get
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
      security:
      - OAuth2PasswordBearer: []
      summary: Get Overall Progress
      tags:
      - progress
  /api/progress/preferences:
    get:
      description: Get user's learning preferences
      operationId: get_user_preferences_api_progress_preferences_get
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPreferencesOut'
          description: Successful Response
      security:
      - OAuth2PasswordBearer: []
      summary: Get User Preferences
      tags:
      - progress
    put:
      description: Update user's learning preferences
      operationId: update_user_preferences_api_progress_preferences_put
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserPreferencesUpdate'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPreferencesOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - OAuth2PasswordBearer: []
      summary: Update User Preferences
      tags:
      - progress
  /api/progress/skills:
    get:
      description: Get progress grouped by skills/categories
      operationId: get_skill_progress_api_progress_skills_get
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
      security:
      - OAuth2PasswordBearer: []
      summary: Get Skill Progress
      tags:
      - progress
  /api/progress/streak:
    post:
      description: Update user's learning streak
      operationId: update_streak_api_progress_streak_post
      parameters:
      - in: query
        name: learning_activity
        required: true
        schema:
          title: Learning Activity
          type: boolean
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - OAuth2PasswordBearer: []
      summary: Update Streak
      tags:
      - progress
...